Implemented .csv file and update_csv module
Created a new module responsible for updating the information on the portugal_data.csv file.
This file will be updated everytime the main script setup.py is run (PDFs will also be downloaded).
I also improved how the get_summary_data() method returns the data: now it does it by returning a dictionary with all the important fields as keys - and not simply as an array.
Soon enough it'll be able to generate the timeline graphs.

Next thing I'll do is to make sure there is no wrong data point being collected from the PDF due to its formatting differences. I already know how the formatting differ for the problem I'm having and that's where I'll begin.