Improved way to access converted text from pdf and implemented csv file update
In the report module, I use the function splitlines() to convert the string generated from the function pdf.convert_pdf_to_txt() to a list of lines. Then I use the same method then before to find the data needed.
When it comes to formatting errors, I managed to deal with a few obvious ones for the gathering of the data (different position of text or data inside the pdf). The solution I used was to go through the items in the list (in the interval selected for the data) and only get the numbers out of them. Since I know the other information are all text of some sort, right now this approuch seems to work. However future formatting errors can occur and I need to pay attention to that.
Right now when the setup.py script is run, it will look into the portugal_data.csv file and check which reports it needs to download to update it to the most recent DGS stance. Then when it downloads all the PDFs and updates the csv file, it then gets the latest DGS report and collects the information to generate the graphs and tables.
Next I'll work on the timeline graphs (since I already have all the data I need in the csv file)